første iteration - Mangler at implementere så den leder efter dynamisk char

+++ src/convert.h
@@ -11,4 +11,4 @@ struct person {
 struct person convert_america(int line_person);
 int Lav_tal_om_for_soren(int tal);
 int convert_borda(int line_person, char* array_pref);
-int welfare_calculator(char winner);
+double welfare_calculator(char winner);
